# src/main/resources/haproxy-swagger.yaml
swagger: '2.0'
info:
  version: 1.0.0
  title: HAProxy Rest API
  description: This is a rest api for haproxy.
  termsOfService: 'http://www.gncloud.kr'
  contact:
    email: support@gncloud.kr
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
host: haproxy.gncloud.io
basePath: /v1
schemes:
  - http
produces:
  - application/json
paths:
...
  /config/global:
    get:
      summary: Get global config.
      description: ''
      operationId: getGlobal
      consumes:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Global'
    post:
      summary: Set global config.
      description: ''
      operationId: setGlobal
      consumes:
        - application/json
      parameters:
        - in: body
          name: global
          description: The global to set.
          required: true
          schema:
            $ref: '#/definitions/Global'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Global'
    put:
      summary: Update global config.
      description: ''
      operationId: updateGlobal
      consumes:
        - application/json
      parameters:
        - in: body
          name: global
          description: The global to update.
          required: true
          schema:
            $ref: '#/definitions/Global'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Global'
  /config/defaults:
    get:
      summary: Get defaults config.
      description: ''
      operationId: getDefaults
      consumes:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Defaults'
    post:
      summary: Set defaults config.
      description: ''
      operationId: setDefaults
      consumes:
        - application/json
      parameters:
        - in: body
          name: defaults
          description: The defaults to set.
          required: true
          schema:
            $ref: '#/definitions/Defaults'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Defaults'

    put:
      summary: Update defaults config.
      description: ''
      operationId: updateDefaults
      consumes:
        - application/json
      parameters:
        - in: body
          name: defaults
          description: The defaults to update.
          required: true
          schema:
            $ref: '#/definitions/Defaults'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Defaults'
  /config/frontends:
    get:
      summary: Get frontend config list.
      description: ''
      operationId: getFrontends
      consumes:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Frontends'
    post:
      summary: Add new frontend.
      description: ''
      operationId: newFrontend
      consumes:
        - application/json
      parameters:
        - in: body
          name: frontend
          description: The frontend to create.
          required: true
          schema:
            $ref: '#/definitions/Frontend'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Frontend'
    put:
      summary: Update frontends config.
      description: ''
      operationId: updateFrontends
      consumes:
        - application/json
      parameters:
        - in: body
          name: body
          description: The frontends to update.
          required: true
          schema:
            $ref: '#/definitions/Frontends'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Frontends'
  '/config/frontends/{frontendId}':
    get:
      summary: Get frontend config.
      description: ''
      operationId: getFrontend
      consumes:
        - application/json
      parameters:
        - name: frontendId
          in: path
          description: ID of frontend to return
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Frontend'
    put:
      summary: Update frontend config.
      description: ''
      operationId: updateFrontend
      consumes:
        - application/json
      parameters:
        - name: frontendId
          in: path
          description: ID of frontend to return
          required: true
          type: string
        - in: body
          name: body
          description: The frontend to update.
          required: true
          schema:
            $ref: '#/definitions/Frontend'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Frontend'
    delete:
      summary: Delete frontend config.
      description: ''
      operationId: deleteFrontend
      parameters:
        - name: frontendId
          in: path
          description: ID of frontend to return
          required: true
          type: string
      responses:
        '200':
          description: OK
  '/config/frontends/{frontendId}/acls':
    get:
      summary: Get acl config list.
      description: ''
      operationId: getAcls
      consumes:
        - application/json
      parameters:
        - name: frontendId
          in: path
          description: ID of frontend to return
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ACLs'
    post:
      summary: Add acl config.
      description: ''
      operationId: newAcl
      consumes:
        - application/json
      parameters:
        - name: frontendId
          in: path
          description: ID of frontend to return
          required: true
          type: string
        - in: body
          name: body
          description: The acl to update.
          required: true
          schema:
            $ref: '#/definitions/ACL'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ACL'
  '/config/frontends/{frontendId}/acls/{aclId}':
    get:
      summary: Get acl config.
      description: ''
      operationId: getAcl
      consumes:
        - application/json
      parameters:
        - name: frontendId
          in: path
          description: ID of frontend to return
          required: true
          type: string
        - name: aclId
          in: path
          description: ID of acl to return
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ACL'
    put:
      summary: Update acl config.
      description: ''
      operationId: updateAcl
      consumes:
        - application/json
      parameters:
        - name: frontendId
          in: path
          description: ID of frontend to return
          required: true
          type: string
        - name: aclId
          in: path
          description: ID of acl to return
          required: true
          type: string
        - in: body
          name: body
          description: The acl to update.
          required: true
          schema:
            $ref: '#/definitions/ACL'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ACL'
    delete:
      summary: Delete acl config.
      description: ''
      operationId: deleteAcl
      parameters:
        - name: frontendId
          in: path
          description: ID of frontend to return
          required: true
          type: string
        - name: aclId
          in: path
          description: ID of acl to return
          required: true
          type: string
      responses:
        '200':
          description: OK
  /config/backends:
    get:
      summary: Get backend config list.
      description: ''
      operationId: getBackends
      consumes:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Backends'
    post:
      summary: Add new backend.
      description: ''
      operationId: newBackend
      consumes:
        - application/json
      parameters:
        - in: body
          name: backend
          description: The backend to create.
          required: true
          schema:
            $ref: '#/definitions/Backend'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Backend'
    put:
      summary: Update backends config.
      description: ''
      operationId: updateBackends
      consumes:
        - application/json
      parameters:
        - in: body
          name: body
          description: The backends to update.
          required: true
          schema:
            $ref: '#/definitions/Backends'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Backends'
  '/config/backends/{backendId}':
    get:
      summary: Get backend config.
      description: ''
      operationId: getBackend
      consumes:
        - application/json
      parameters:
        - name: backendId
          in: path
          description: ID of backend
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Backend'
    put:
      summary: Update backend config.
      description: ''
      operationId: updateBackend
      consumes:
        - application/json
      parameters:
        - name: backendId
          in: path
          description: ID of backend
          required: true
          type: string
        - in: body
          name: body
          description: The backend to update.
          required: true
          schema:
            $ref: '#/definitions/Backend'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Backend'
    delete:
      summary: Delete backend config.
      description: ''
      operationId: deleteBackend
      parameters:
        - name: backendId
          in: path
          description: ID of backend
          required: true
          type: string
      responses:
        '200':
          description: OK
  '/config/backends/{backendId}/servers':
    get:
      summary: Get server config list.
      description: ''
      operationId: getServers
      consumes:
        - application/json
      parameters:
        - name: backendId
          in: path
          description: ID of backend
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Servers'
    post:
      summary: Add server config.
      description: ''
      operationId: newServer
      consumes:
        - application/json
      parameters:
        - name: backendId
          in: path
          description: ID of backend
          required: true
          type: string
        - in: body
          name: body
          description: The server to add.
          required: true
          schema:
            $ref: '#/definitions/Server'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Server'
  '/config/backends/{backendId}/servers/{serverId}':
    get:
      summary: Get server config.
      description: ''
      operationId: getServer
      consumes:
        - application/json
      parameters:
        - name: backendId
          in: path
          description: ID of backend
          required: true
          type: string
        - name: serverId
          in: path
          description: ID of server
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Server'
    put:
      summary: Update server config.
      description: ''
      operationId: updateServer
      consumes:
        - application/json
      parameters:
        - name: backendId
          in: path
          description: ID of backend
          required: true
          type: string
        - name: serverId
          in: path
          description: ID of server
          required: true
          type: string
        - in: body
          name: body
          description: The backend to update.
          required: true
          schema:
            $ref: '#/definitions/Server'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Server'
    delete:
      summary: Delete server config.
      description: ''
      operationId: deleteServer
      parameters:
        - name: backendId
          in: path
          description: ID of backend
          required: true
          type: string
        - name: serverId
          in: path
          description: ID of server
          required: true
          type: string
      responses:
        '200':
          description: OK
definitions:
...
  Global:
    type: object
    properties:
      name:
        type: string
      mode:
        type: string
      maxconn:
        type: integer
        format: int64
    additionalProperties:
      $ref: '#/definitions/Options'
  Defaults:
    type: object
    properties:
      mode:
        type: string
      timeoutConnect:
        type: string
      timeoutClient:
        type: string
      timeoutServer:
        type: string
    additionalProperties:
      $ref: '#/definitions/Options'
  Frontends:
    type: object
    additionalProperties:
      $ref: '#/definitions/Frontend'
  Frontend:
    type: object
    properties:
      name:
        type: string
      mode:
        type: string
      bindPort:
        type: integer
        format: int64
      bindIp:
        type: string
      defaultBackend:
        type: string
      timeoutConnect:
        type: string
      timeoutClient:
        type: string
      timeoutServer:
        type: string
      acls:
        $ref: '#/definitions/ACLs'
      httpSpikeLimit:
        $ref: '#/definitions/SpikeLimit'
      tcpSpikeLimit:
        $ref: '#/definitions/SpikeLimit'
    additionalProperties:
      $ref: '#/definitions/Options'
  ACLs:
      type: object
      additionalProperties:
        $ref: '#/definitions/ACL'
  ACL:
    type: object
    properties:
      name:
        type: string
      backend:
        type: string
      pattern:
        type: string
  SpikeLimit:
    type: object
    properties:
      sampleTime:
        type: string
      expiryTime:
        type: string
      rate:
        type: integer
        format: int64
  Backends:
    type: object
    additionalProperties:
      $ref: '#/definitions/Backend'
  Backend:
    type: object
    properties:
      name:
        type: string
      mode:
        type: string
      servers:
        $ref: '#/definitions/Servers'
    additionalProperties:
      $ref: '#/definitions/Options'
  Servers:
    type: object
    additionalProperties:
      $ref: '#/definitions/Server'
  Server:
    type: object
    properties:
      name:
        type: string
      host:
        type: string
      port:
        type: integer
        format: int64
      weight:
        type: integer
        format: int64
      maxconn:
        type: integer
        format: int64
      check:
        type: boolean
      checkInterval:
        type: integer
        format: int64
  Options:
    type: array
    items:
      type: object